ecommerce/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-auditlog==3.0.0
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
filelock==3.14.0
idna==3.7
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
py3-validate-email==1.0.5.post1
py3dns==4.0.1
PyJWT==1.7.1
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
validate-email==1.3
validate-email-address==1
